requirements.txt
Flask==3.0.2
python-dotenv==1.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
google-generativeai==0.8.3
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0
